added wmin api and parse_wminpdfset as wmin parse rule
